+++ src/projects/NotBopIt.tsx
@@ -179,7 +179,6 @@ export const NotBopIt = {
   },
   description: "Bop It wishes it was this",
   image: "/assets/project_images/NotBopIt/microbitLogo.png",
-  tags: ["Sensors", "Embedded", "Micro:bit"],
+  tags: ["C++", "Sensors", "Microcontroller Software"],
   body: <BopIt />,
 };
-