+++ src/projects/SpeedReader.tsx
@@ -9,7 +9,6 @@ export const SpeedReader = {
   },
   description: "Read 3x faster",
   image: "/assets/project_images/SpeedReader/sr.png",
-  tags: ["OCR", "React", "Web"],
-  link: "https://www.speedread.dev/"
+  tags: ["Next.js", "Supabase"],
+  link: "https://www.speedread.dev/",
 };
-