+++ src/projects/Verdra.tsx
@@ -9,7 +9,6 @@ export const Verdra = {
   },
   description: "Optimize your cloud computing",
   image: "/assets/project_images/Verdra/verdra.png",
-  tags: ["Cloud", "Optimization", "Web"],
+  tags: ["Python", "AWS Lambda", "AST Parsing"],
   link: "https://verdratech.github.io/Verdra-Site/",
 };
-