+++ CHANGELOG.md
@@ -165,3 +165,4 @@
 
 - Redesigned projects and music page
 - Redesigned projects and music cards
+- Added skills to project cards